+++ src/mruby_integration/models/music.cpp
@@ -160,6 +160,17 @@ mrb_Music_stop(mrb_state* mrb, mrb_value self) -> mrb_value
   return mrb_nil_value();
 }
 
+auto
+mrb_Music_played(mrb_state* mrb, mrb_value self) -> mrb_value
+{
+  Music* music;
+
+  Data_Get_Struct(mrb, self, &Music_type, music);
+  mrb_assert(music != nullptr);
+
+  return mrb_float_value(mrb, GetMusicTimePlayed(*music));
+}
+
 void
 append_models_Music(mrb_state* mrb)
 {
@@ -177,6 +188,8 @@ append_models_Music(mrb_state* mrb)
   mrb_define_method(
     mrb, Music_class, "playing?", mrb_Music_playing, MRB_ARGS_NONE());
   mrb_define_method(mrb, Music_class, "stop", mrb_Music_stop, MRB_ARGS_NONE());
+  mrb_define_method(
+    mrb, Music_class, "played", mrb_Music_played, MRB_ARGS_NONE());
 
   load_ruby_models_music(mrb);
 }